Introduce root entity props in the space view blueprints and refactor to use this instead of the ad hoc state in the timeseries space view.
